Rename vendored nlohmann-json to avoid other-Linux collisions
