+++ perl/lib/Nix/Store.xs
@@ -15,7 +15,7 @@
 #include "crypto.hh"
 
 #include <sodium.h>
-#include <nlohmann/json.hpp>
+#include <nix-nlohmann/json.hpp>
 
 
 using namespace nix;
