+++ src/libexpr/flake/lockfile.hh
@@ -2,7 +2,7 @@
 
 #include "flakeref.hh"
 
-#include <nlohmann/json_fwd.hpp>
+#include <nix-nlohmann/json_fwd.hpp>
 
 namespace nix {
 class Store;
