+++ src/libfetchers/attrs.hh
@@ -4,7 +4,7 @@
 
 #include <variant>
 
-#include <nlohmann/json_fwd.hpp>
+#include <nix-nlohmann/json_fwd.hpp>
 
 #include <optional>
