+++ src/libfetchers/fetchers.cc
@@ -1,7 +1,7 @@
 #include "fetchers.hh"
 #include "store-api.hh"
 
-#include <nlohmann/json.hpp>
+#include <nix-nlohmann/json.hpp>
 
 namespace nix::fetchers {
