+++ src/libstore/binary-cache-store.cc
@@ -18,7 +18,7 @@
 #include <regex>
 #include <fstream>
 
-#include <nlohmann/json.hpp>
+#include <nix-nlohmann/json.hpp>
 
 namespace nix {
