+++ src/libstore/derived-path.cc
@@ -1,7 +1,7 @@
 #include "derived-path.hh"
 #include "store-api.hh"
 
-#include <nlohmann/json.hpp>
+#include <nix-nlohmann/json.hpp>
 
 namespace nix {
