+++ src/libstore/globals.cc
@@ -11,7 +11,7 @@
 #include <dlfcn.h>
 #include <sys/utsname.h>
 
-#include <nlohmann/json.hpp>
+#include <nix-nlohmann/json.hpp>
 
 
 namespace nix {
