+++ src/libstore/nar-accessor.cc
@@ -6,7 +6,7 @@
 #include <stack>
 #include <algorithm>
 
-#include <nlohmann/json.hpp>
+#include <nix-nlohmann/json.hpp>
 
 namespace nix {
