+++ src/libstore/parsed-derivations.hh
@@ -2,7 +2,7 @@
 
 #include "store-api.hh"
 
-#include <nlohmann/json_fwd.hpp>
+#include <nix-nlohmann/json_fwd.hpp>
 
 namespace nix {
