+++ src/libutil/args.cc
@@ -3,7 +3,7 @@
 
 #include <glob.h>
 
-#include <nlohmann/json.hpp>
+#include <nix-nlohmann/json.hpp>
 
 namespace nix {
