+++ src/libutil/args.hh
@@ -4,7 +4,7 @@
 #include <map>
 #include <memory>
 
-#include <nlohmann/json_fwd.hpp>
+#include <nix-nlohmann/json_fwd.hpp>
 
 #include "util.hh"
