+++ src/nix-nlohmann/local.mk
@@ -1,0 +1,2 @@
+$(foreach i, $(wildcard src/nix-nlohmann/*.hpp), \
+  $(eval $(call install-file-in, $(i), $(includedir)/nix-nlohmann, 0644)))
