+++ src/nix/develop.cc
@@ -9,7 +9,7 @@
 #include "run.hh"
 
 #include <memory>
-#include <nlohmann/json.hpp>
+#include <nix-nlohmann/json.hpp>
 
 using namespace nix;
