+++ src/nix/main.cc
@@ -18,7 +18,7 @@
 #include <netdb.h>
 #include <netinet/in.h>
 
-#include <nlohmann/json.hpp>
+#include <nix-nlohmann/json.hpp>
 
 extern std::string chrootHelperName;
