+++ src/nix/profile.cc
@@ -10,7 +10,7 @@
 #include "profiles.hh"
 #include "names.hh"
 
-#include <nlohmann/json.hpp>
+#include <nix-nlohmann/json.hpp>
 #include <regex>
 #include <iomanip>
