--- src/nlohmann/local.mk
@@ -1,2 +1,0 @@
-$(foreach i, $(wildcard src/nlohmann/*.hpp), \
-  $(eval $(call install-file-in, $(i), $(includedir)/nlohmann, 0644)))
